{{ cookiecutter.project_slug }}/requirements.txt
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.4.0
pyyaml>=6.0.0

pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0
mypy>=1.10.0